Refactored thread storage; removed the base class; migrated to pymongo (motor will be deprecated)
